requirements.txt
alembic==1.3.3
aniso8601==8.0.0
argcomplete==1.11.1
attrs==19.3.0
boto3==1.11.12
botocore==1.14.12
certifi==2019.11.28
cfn-flip==1.2.2
chardet==3.0.4
Click==7.0
coverage==5.0.3
docutils==0.15.2
durationpy==0.5
Flask==1.1.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
future==0.18.2
hjson==3.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
jsonschema==3.2.0
kappa==0.6.0
lambda-packages==0.20.0
Mako==1.1.1
MarkupSafe==1.1.1
placebo==0.9.0
psycopg2-binary==2.8.4
pyrsistent==0.15.7
python-dateutil==2.6.1
python-dotenv==0.10.5
python-editor==1.0.4
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.13
text-unidecode==1.3
toml==0.10.0
tqdm==4.42.1
troposphere==2.5.3
urllib3==1.25.8
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.49.0
